ChaoticNDETools

Documentation for ChaoticNDETools.

ChaoticNDETools.NablaSkipConnection

ChaoticNDETools.ParSkipConnection

ChaoticNDETools.average_forecast_length

ChaoticNDETools.forecast_δ

ChaoticNDETools.gpuoff

ChaoticNDETools.gpuon

ChaoticNDETools.∂x_PBC

ChaoticNDETools.∂x²_PBC

ChaoticNDETools.∂x⁴_PBC

ChaoticNDETools.NablaSkipConnection — Type

NablaSkipConnection

With the input $x$, passes $w * (∇ * x) + (1 - |w|) * x$ , where $w\in\mathbb{R}, w\in [0,1]$. Here $\Nabla$ is a finite difference derivative matrix

ChaoticNDETools.ParSkipConnection — Type

ParSkipConnection(layer, connection)

A version of SkipConnection that is paramaterized with one single parameter, thus the output is connection(w .* layer(input), input)

ChaoticNDETools.average_forecast_length — Function

average_forecast_length(predict, valid::NODEDataloader,N_t=300; λmax=0, mode="norm")

Returns the average forecast length on a NODEDataloader set (should be valid or test set) given a (t, u0) -> prediction function. N_t is the length of each forecast, has to be larger than the expected forecast length. If a λmax is given, the results are scaled with it (and dt`)

ChaoticNDETools.forecast_δ — Method

forecast_δ(prediction::AbstractArray{T,N}, truth::AbstractArray{T,N}, mode::String="both") where {T,N}

Assumes that the last dimension of the input arrays is the time dimension and N_t long. Returns an N_t long array, judging how accurate the prediction is.

Supported modes:

"mean": mean between the arrays

"maximum": maximum norm

"norm": normalized, similar to the metric used in Pathak et al

ChaoticNDETools.gpuoff — Method

gpuoff()

Manually toggle GPU use off

ChaoticNDETools.gpuon — Method

gpuon()

Manually toggle GPU use on (if available)

ChaoticNDETools.∂x_PBC — Method

∂x_PBC(n::Integer, dx::T)

2nd order central finite difference matrix for 1d domains with periodic boundary conditions

ChaoticNDETools.∂x²_PBC — Method

∂x²_PBC(n::Integer, dx::T)

2nd order central finite difference matrix of the second derivative for 1d domains with periodic boundary conditions

ChaoticNDETools.∂x⁴_PBC — Method

∂x⁴_PBC(n::Integer, dx::T)

4th derivative finite difference matrix with periodic boundary conditions